pyproject.toml
[project]
name = "refu"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "matplotlib>=3.10.1",
    "pytelegrambotapi>=4.26.0",
    "python-dotenv>=1.1.0",
    "sqlalchemy>=2.0.40",
]
